requirements.txt
fastapi
uvicorn[standard]
stripe
resend
python-dotenv
